+++ src/main/java/com/nhnacademy/midoo/gateway/domain/CommentPostRequest.java
@@ -5,6 +5,6 @@ import lombok.Value;
 @Value
 public class CommentPostRequest {
     String commentContent;
-    int taskId;
-    int accountId;
+    long taskId;
+    long accountId;
 }
